refactor: change file name to historyConfig
